requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.3
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.0
drf-yasg==1.20.0
environs==9.3.2
gunicorn==20.1.0
h11==0.12.0
httptools==0.1.2
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
marshmallow==3.12.1
packaging==20.9
Pillow==8.2.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
pyparsing==2.4.7
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
